Create student-class-wheel.html
Sizing works, but what a pain. Would prefer if sizing were based on percentage instead of an absolute number (and children's totals have to be less than the parent... all of which has to be manually calculated).

Next up: fix color to be explicitly specified in the JSON instead the current system (which I don't understand... [2] seems to be used both for color and for other mysterious things. But I don't understand the color function. Is it an RGB look-up?